64-bit types must be `long long` thanks @Sauraen!

+++ stdint.h
@@ -9,13 +9,13 @@
 typedef signed char          int8_t;
 typedef signed short int     int16_t;
 typedef signed int           int32_t;
-typedef signed long int      int64_t;
+typedef signed long long     int64_t;
 
 /* Unsigned.  */
 typedef unsigned char        uint8_t;
 typedef unsigned short int   uint16_t;
 typedef unsigned int         uint32_t;
-typedef unsigned long int    uint64_t;
+typedef unsigned long long   uint64_t;
 
 typedef __SIZE_TYPE__        size_t;
